match: SOURCE/CURSOR and SOURCE/FINDPATH -> 100%, PATH 14/18
All 18 assigned functions exact. CURSOR 20/20 and FINDPATH 14/14, both units
at 100.00%.

ProcessMapChange went 0.00 -> EXACT on the span alone: VA(0x00435ad8) was 0xb2
where the function actually runs to 0x705, with Ghidra cutting before an
11-entry jump table. MoveHero likewise gains 0x44 for its dword table and
0x3c-byte case-index table, ending exactly at CheckAdjacentMon. FindCombatPath
went 47.72 -> EXACT on a full rebuild.

The `_$E` static-initializer chain names the TU that owns a global. Decoding
PATH's four funclets byte by byte - _$E17 calls _$E1 and _$E18; _$E1 constructs
SVSearchArray; _$E18 registers the dtor with _atexit - proves retail's PATH
translation unit defines `searchArray SVSearchArray;`, where our tree has it in
PHILAI.cpp:404. The lane measured the move (PATH 14/18 -> 16/18, PHILAI
unchanged, no regressions) and reverted it as out of scope; it is worth doing
deliberately. Written up as static-init-funclet-chain-names-owner-tu.md.

It also settled an open ARMY question: army::ValidAttack writes its out-param
exactly once at offset 0 - every store is `mov [ecx],eax`, there is no [ecx+4]
anywhere, and GetAttackMask passes a plain 4-byte slot - so the contract is a
single i32 and AttemptAdjacentAttack's second slot is its own.

Other recoveries: CalcTerrainCost has no locals at all (its frame is just the
two spilled fastcall params); SeedCombatPosition needs seven dead locals to
reach retail's 0x2c frame and gates on m_monster.shots, not speed;
TestPossibleDirections spells thirteen longhand `gSearchTerrain = INVALID;
goto storeDirection;` where we had a loop; EveryOther is u8, not i32.

Jump stubs are parked in REVERSE source order, which turns a stub block into a
goto census - that is how MoveHero's two missing `goto movementDone;` were
found. Appended to goto-label-jump-stub.md.

Tree 1473 -> 1491, code bytes 64.46% -> 66.12%, fuzzy 94.51% -> 94.77%. No
regressions.

+++ include/SOURCE/advManager.h
@@ -380,7 +380,7 @@ extern i32 giPixelsPerStep[ADVMGR_STEP_PIXEL_COUNT];
 extern i32 giStepDelay[ADVMGR_STEP_DELAY_COUNT];
 extern class sample* hOldWalkSample;
 extern class sample* hNewWalkSample;
-extern i32 EveryOther;
+extern u8 EveryOther;
 extern i32 startVals[ADVMGR_VIEW_WORLD_SCALE_COUNT];
 extern i8 iVWHalf[ADVMGR_VIEW_WORLD_SCALE_COUNT][ADVMGR_VIEW_WORLD_OFFSET_KIND_COUNT]
                      [IDX(COORDINATE_AXIS_COUNT)];
